requirements.txt
annotated-types
contourpy
cycler
faiss-cpu
fonttools
imageio
joblib
kiwisolver
lazy-loader
matplotlib
neo4j
networkx
opencv-contrib-python-headless
opencv-contrib-python
opencv-python
packaging
pandas
pillow
pydantic-core
pydantic
pyparsing
python-dateutil
pytz
scikit-image
scikit-learn
scipy
six
threadpoolctl
tifffile
typing-extensions
tzdata
seaborn
tensorflow
psutil